version 3.2.1

+++ application/models/Tasks_model.php
@@ -1205,12 +1205,14 @@ class Tasks_model extends App_Model
             if (empty($attachment->external)) {
                 $relPath  = get_upload_path_by_type('task') . $attachment->rel_id . '/';
                 $fullPath = $relPath . $attachment->file_name;
-                unlink($fullPath);
-                $fname     = pathinfo($fullPath, PATHINFO_FILENAME);
-                $fext      = pathinfo($fullPath, PATHINFO_EXTENSION);
-                $thumbPath = $relPath . $fname . '_thumb.' . $fext;
-                if (file_exists($thumbPath)) {
-                    unlink($thumbPath);
+                if(file_exists($fullPath)) {
+                    unlink($fullPath);
+                    $fname     = pathinfo($fullPath, PATHINFO_FILENAME);
+                    $fext      = pathinfo($fullPath, PATHINFO_EXTENSION);
+                    $thumbPath = $relPath . $fname . '_thumb.' . $fext;
+                    if (file_exists($thumbPath)) {
+                        unlink($thumbPath);
+                    }
                 }
             }
